+++ src/burn/engine/core.cpp
@@ -527,7 +527,7 @@ extern "C" HRESULT CoreApply(
     BOOL fSuspend = FALSE;
     BOOTSTRAPPER_APPLY_RESTART restart = BOOTSTRAPPER_APPLY_RESTART_NONE;
     BURN_CACHE_THREAD_CONTEXT cacheThreadContext = { };
-    LONGLONG llNumberOfPhases = 0;
+    DWORD dwNumberOfPhases = 0;
 
     LogId(REPORT_STANDARD, MSG_APPLY_BEGIN);
 
@@ -539,14 +539,13 @@ extern "C" HRESULT CoreApply(
 
     if (pEngineState->plan.cCacheActions)
     {
-        ++llNumberOfPhases;
+        ++dwNumberOfPhases;
     }
     if (pEngineState->plan.cExecuteActions)
     {
-        ++llNumberOfPhases;
+        ++dwNumberOfPhases;
     }
-    hr = VariableSetNumeric(&pEngineState->variables, BURN_BUNDLE_NUMBER_OF_APPLY_PHASES, llNumberOfPhases, TRUE);
-    //don't fail - best effort only
+    pEngineState->userExperience.pUserExperience->OnApplyNumberOfPhases(dwNumberOfPhases);
 
     int nResult = pEngineState->userExperience.pUserExperience->OnApplyBegin();
     hr = UserExperienceInterpretResult(&pEngineState->userExperience, MB_OKCANCEL, nResult);